+++ xray/entrypoint.sh
@@ -10,7 +10,9 @@ for client in $XRAY_CLIENTS; do
     *)   id="$client" ; email="" ;;
   esac
   [ -n "$CLIENTS_JSON" ] && CLIENTS_JSON="$CLIENTS_JSON,"
-  CLIENTS_JSON="$CLIENTS_JSON{\"id\":\"$id\",\"flow\":\"xtls-rprx-vision\""
+  FLOW_STR="xtls-rprx-vision"
+  [ "${XRAY_NETWORK:-tcp}" = "xhttp" ] && FLOW_STR=""
+  CLIENTS_JSON="$CLIENTS_JSON{\"id\":\"$id\",\"flow\":\"$FLOW_STR\""
   [ -n "$email" ] && CLIENTS_JSON="$CLIENTS_JSON,\"email\":\"$email\""
   CLIENTS_JSON="$CLIENTS_JSON}"
 done
@@ -43,8 +45,32 @@ CLIENTS_JSON="$CLIENTS_JSON" NAMES_JSON="$NAMES_JSON" SIDS_JSON="$SIDS_JSON" \
   XRAY_LOG_LEVEL="$XRAY_LOG_LEVEL" \
   XRAY_REALITY_DEST="$XRAY_REALITY_DEST" \
   XRAY_REALITY_PRIVATE_KEY="$XRAY_REALITY_PRIVATE_KEY" \
+  XRAY_NETWORK="${XRAY_NETWORK:-tcp}" \
+  XRAY_XHTTP_MODE="${XRAY_XHTTP_MODE:-auto}" \
+  XRAY_XHTTP_PATH="${XRAY_XHTTP_PATH:-/}" \
   python3 -c '
 import json, os
+network = os.environ["XRAY_NETWORK"]
+stream_settings = {
+  "network": network,
+  "security": "reality",
+  "realitySettings": {
+    "show": False,
+    "dest": os.environ["XRAY_REALITY_DEST"],
+    "xver": 0,
+    "serverNames": json.loads("[" + os.environ["NAMES_JSON"] + "]"),
+    "privateKey": os.environ["XRAY_REALITY_PRIVATE_KEY"],
+    "shortIds": json.loads("[" + os.environ["SIDS_JSON"] + "]")
+  },
+  "packetEncoding": "xudp"
+}
+if network == "xhttp":
+  stream_settings["xhttpSettings"] = {
+    "mode": os.environ["XRAY_XHTTP_MODE"],
+    "path": os.environ["XRAY_XHTTP_PATH"]
+  }
+else:
+  stream_settings["tcpSettings"] = {"header": {"type": "none"}}
 config = {
   "log": {
     "loglevel": os.environ["XRAY_LOG_LEVEL"],
@@ -60,20 +86,7 @@ config = {
       "clients": json.loads("[" + os.environ["CLIENTS_JSON"] + "]"),
       "decryption": "none"
     },
-    "streamSettings": {
-      "network": "tcp",
-      "security": "reality",
-      "realitySettings": {
-        "show": False,
-        "dest": os.environ["XRAY_REALITY_DEST"],
-        "xver": 0,
-        "serverNames": json.loads("[" + os.environ["NAMES_JSON"] + "]"),
-        "privateKey": os.environ["XRAY_REALITY_PRIVATE_KEY"],
-        "shortIds": json.loads("[" + os.environ["SIDS_JSON"] + "]")
-      },
-      "tcpSettings": {"header": {"type": "none"}},
-      "packetEncoding": "xudp"
-    },
+    "streamSettings": stream_settings,
     # --sniffing
     "sniffing": {
       "enabled": True,